refactor: extract manager SPA routes from main into dedicated router
